+++ src/models/language-keywords.ts
@@ -1,0 +1,16 @@
+"use strict";
+
+//#region Language keywords
+export const KEYWORDS = new Set([
+	'if', 'else', 'while', 'repeat', 'for', 'in', 'break', 'continue',
+	'return', 'use', 'from', 'this', 'true', 'false', 'null'
+]);
+
+export const BUILT_IN_TYPES = new Set([
+	'Number', 'String', 'Boolean', 'Any'
+]);
+
+export const BUILT_IN_FUNCTIONS = new Set([
+	'write', 'read'
+]);
+//#endregion